Options for widget's iframe.

Taken from http://blog.apps.npr.org/pym.js/api/pym.js/1.1.2/module-pym.Parent.html. In addition also check out the https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe.

frameOptions(xdomain = "*", title = NULL, name = NULL, id = NULL,
  allowfullscreen = FALSE, sandbox = NULL)

Arguments

xdomain

xdomain to validate messages received.

title

If passed it will be assigned to the iframe title attribute.

name

If passed it will be assigned to the iframe name attribute.

id

If passed it will be assigned to the iframe id attribute.

allowfullscreen

If TRUE it will set the iframe allowfullscreen attribute to true.

sandbox

If passed it will be assigned to the iframe sandbox attribute.