Save a widgetframe and its child widget to HTML files.

Similar to saveWidget() with the addition that both the parent widget and the enclosed child widget are saved to two different HTML files.

saveWidgetframe(widget, file, selfcontained = FALSE, libdir = NULL,
  background = "white", knitrOptions = list())

Arguments

widget

widgetframe to save

file

File to save the parent widget into. The child widget will be saved to `basename(file)_widget/index.html`.

selfcontained

Whether to save the parent and child HTMLs as a single self-contained files. WARNING: Setting this option to true will still result in two HTMLs, one for the parent and another for the child widget. (with external resources base64 encoded) or files with external resources placed in an adjacent directory.

libdir

Directory to copy HTML dependencies into (defaults to filename_files).

background

Text string giving the html background color of the widget. Defaults to white.

knitrOptions

A list of knitr chunk options.